move location of a class
